add tinymce plugins editor

+++ app/Http/Controllers/Dashboard/ServicesController.php
@@ -86,7 +86,7 @@ class ServicesController extends Controller
      * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
      */
 
-    public function edit($id)
+    public function edit(Service $service,$id)
     {
         return view('dashboard.services.edit')->with([
             'service' => Service::find($id),
@@ -174,17 +174,22 @@ class ServicesController extends Controller
         return redirect()->route('service.index');
     }
 
-    public function removeFile($id) {
+    public function removeFile($id){
 
         $service = $this->repository->pushCriteria(SelfServiceCriteria::class)->find($id);
+
+        if ($service->file == null) {
+            return "error";
+        }
 
         if (\Storage::disk('public')->exists($service->file)) :
             \Storage::disk('public')->delete($service->file);
         endif;
 
         if ($service->update(['file' => null,])) {
-            return response('ok');
+            return 'ok';
         }
+
     }
 
 }
